+++ src/app/data/db.service.ts
@@ -62,8 +62,8 @@ export class DbService {
     return until;
   }
 
-  public async findEvents(query: string, day: Date | undefined, category: string, coords: GpsCoord | undefined, timeRange: TimeRange | undefined): Promise<Event[]> {
-    return await call(this.worker, DataMethods.FindEvents, query, day, category, coords, timeRange);
+  public async findEvents(query: string, day: Date | undefined, category: string, coords: GpsCoord | undefined, timeRange: TimeRange | undefined, allDay: boolean): Promise<Event[]> {
+    return await call(this.worker, DataMethods.FindEvents, query, day, category, coords, timeRange, allDay);
   }
 
   public async getEventList(ids: string[]): Promise<Event[]> {